requirements.txt
boto3==1.9.134
botocore==1.12.134
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
dynamodb-json==1.3
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
py-trello==0.14.0
python-dateutil==2.8.0
python-lambda==3.2.6
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
urllib3==1.24.2
